requirements.txt
alabaster==0.7.13
asgiref==3.7.2
Babel==2.13.0
bootstrap-py==2.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.6
docutils==0.20.1
gitdb==4.0.10
GitPython==3.1.37
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pguard==0.4.0
Pillow==10.0.1
Pygments==2.16.1
requests==2.31.0
setuptools==68.2.2
setuptools-scm==8.0.4
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
